boot: Add/Update logs during FIT signature verification
Move log wrt signature node in FIT being verified for config part to a
more appropriate place i.e., after ensuring node is a signature node.

Add logs wrt signature node in FIT being verified for image part as well.
The logic of the default code is a bit different and hence logs may feel
to be a bit confusing.
  - All the keys used for all the various images gets collected in the
    public key DTB.
  - The code loops all keys (one by one) over one individual image at
    a time even if a certain key was not used for signature over that
    particular image.
  - The signature check for each key is actually performed against the
    key that is reflected by the FIT and not against the key because of
    which the loop has happened.
  - All this explains the reason of same log with +(means pass) getting
    displayed for each key.

Signed-off-by: Ranbir Singh <rsingh@ventanamicro.com>